Rename AeTextArea to AeTextarea for consistency with Textarea element
